[refactor] rename to correct package
