+++ app/src/main/java/com/mashup/zuzu/ui/leave/Leave.kt
@@ -1,4 +1,4 @@
-package com.mashup.zuzu.ui.user
+package com.mashup.zuzu.ui.leave
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -16,10 +16,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.mashup.zuzu.ui.component.Button
-import com.mashup.zuzu.ui.leave.LeaveUiEventState
-import com.mashup.zuzu.ui.leave.LeaveViewModel
 import com.mashup.zuzu.ui.theme.ProofTheme
 
 /**
