+++ app/src/main/java/com/mashup/zuzu/ui/leave/LeaveViewModel.kt
@@ -6,7 +6,6 @@ import com.mashup.zuzu.data.model.Results
 import com.mashup.zuzu.domain.usecase.LeaveMembershipUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
